Change typedef tThreadId from uint to ulong to match actual Win32 type for thread ids (DWORD)

+++ tech/include/thread.h
@@ -17,7 +17,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 
 #ifdef _WIN32
-typedef uint tThreadId;
+typedef ulong tThreadId;
 #else
 typedef pthread_t tThreadId;
 #endif
